// MCP_Frontend-main/src/app/shared/models/efris-commodity.ts
export class EfrisCommodity {
     commodityId: number;
     commodityCategoryCode: String;
     commodityCategoryLevel: String;
     commodityCategoryName: String;
     parentCode: String;
}

export class EfrisCommodityCode {
    commodityCategoryLevel1Code: String;
    commodityCategoryLevel2Code: String;
    commodityCategoryLevel3Code: String;
    commodityCategoryLevel4Code: String;
    commodityCategoryLevel5Code: String;
}