// MCP_Frontend-main/src/app/shared/models/goods-data.ts
export class GoodsData {
    operationType: string;
    goodsName: string;
    goodsCode: string;
    measureUnit: string;
    unitPrice: string;
    currency: string;
    commodityCategoryId: string;
    haveExciseTax: string;
    description: string;
    stockPrewarning: string;
    pieceMeasureUnit: string;
    havePieceUnit: string;
    pieceUnitPrice: string;
    packageScaledValue: string;
    pieceScaledValue: string;
    exciseDutyCode: string;
    haveOtherUnit: string;
    commodityCategoryName: string;
    pageNo: string;
    pageSize: string;
    branchId: string;
    id: String;
    stock: String;
    source: String;
    statusCode: String;
    commodityCategoryCode: String;
    taxRate: number;
    isZeroRate: String;
    isExempt: String;
    exciseDutyName: String;
    exciseRate: String;
    pack: String;
    stick: String;
    remarks: String;
    exclusion: String;





}